[Update] theme image option
