+++ src/components/OrderingProducts/SiteTheme/ThemeOption.js
@@ -1,19 +1,14 @@
-import React, { useState, useRef } from 'react'
-import { useLanguage, DragAndDrop, ExamineClick, useToast, ToastType } from 'ordering-components-admin'
-import Skeleton from 'react-loading-skeleton'
-import { ColorPicker, Alert, ImageCrop, Modal } from '../../Shared'
-import BsCardImage from '@meronex/icons/bs/BsCardImage'
-import { bytesConverter } from '../../../utils'
+import React from 'react'
+import { useLanguage } from 'ordering-components-admin'
+import { ColorPicker } from '../../Shared'
 import { SecondSelect as Select, Checkbox, Input } from '../../../styles'
+import { ThemeImage } from './ThemeImage'
 
 import {
   OptionContainer,
   OptionHeader,
   Option,
-  ColorPickerContainer,
-  ComponentImage,
-  UploadImageIconContainer,
-  UploadImageIcon
+  ColorPickerContainer
 } from './styles'
 
 export const ThemeOption = (props) => {
@@ -28,12 +23,6 @@ export const ThemeOption = (props) => {
   } = props
 
   const [, t] = useLanguage()
-  const [, { showToast }] = useToast()
-
-  const imageRef = useRef(null)
-  const [alertState, setAlertState] = useState({ open: false, content: [] })
-  const [cropState, setCropState] = useState({ name: null, data: null, open: false })
-  const [imageUploadState, setImageUploadState] = useState({ loading: false, image: null })
 
   const getOptions = (options) => {
     return options.map(option => {
@@ -66,71 +55,6 @@ export const ThemeOption = (props) => {
     const _themeValues = { ...themeValues }
     updateObject(_themeValues, value, path)
     setThemeValues(_themeValues)
-  }
-
-  const handleClickImage = () => {
-    imageRef.current.click()
-  }
-
-  const handleFiles = (files) => {
-    if (files.length === 1) {
-      const type = files[0].type.split('/')[0]
-      if (type !== 'image') {
-        setAlertState({
-          open: true,
-          content: [t('ERROR_ONLY_IMAGES', 'Only images can be accepted')]
-        })
-        return
-      }
-
-      if (bytesConverter(files[0]?.size) > 2048) {
-        setAlertState({
-          open: true,
-          content: [t('IMAGE_MAXIMUM_SIZE', 'The maximum image size is 2 megabytes')]
-        })
-        return
-      }
-      const reader = new window.FileReader()
-      reader.readAsDataURL(files[0])
-      reader.onload = () => {
-        setCropState({ name: 'photo', data: reader.result, open: true })
-      }
-      reader.onerror = error => console.log(error)
-    }
-  }
-
-  const handleChangePhoto = async (croppedImg) => {
-    try {
-      setImageUploadState({
-        ...imageUploadState,
-        loading: true
-      })
-      showToast(ToastType.Info, t('LOADING', 'Loading'))
-      setCropState({ name: null, data: null, open: false })
-      const { error, result } = await handleAddThemeGallery(croppedImg)
-      if (!error) {
-        showToast(ToastType.Success, t('IMAGE_SAVED', 'Image saved'))
-        handleChangeValue(result?.image)
-      } else {
-        setAlertState({
-          open: true,
-          content: result
-        })
-      }
-      setImageUploadState({
-        loading: false,
-        image: error ? null : result?.image
-      })
-    } catch (error) {
-      setImageUploadState({
-        ...imageUploadState,
-        loading: false
-      })
-      setAlertState({
-        open: true,
-        content: [error.message]
-      })
-    }
   }
 
   return (
@@ -167,36 +91,12 @@ export const ThemeOption = (props) => {
         />
       )}
       {(name === 'image' || name === 'dummy_image') && optionObject?.value_type === 'string' && (
-        <ComponentImage
-          isThemeOption
-          onClick={() => handleClickImage()}
-        >
-          <ExamineClick
-            onFiles={files => handleFiles(files)}
-            childRef={(e) => { imageRef.current = e }}
-            accept='image/png, image/jpeg, image/jpg'
-            disabled={imageUploadState.loading}
-          >
-            <DragAndDrop
-              onDrop={dataTransfer => handleFiles(dataTransfer.files)}
-              accept='image/png, image/jpeg, image/jpg'
-              disabled={imageUploadState.loading}
-            >
-              {imageUploadState.loading && (
-                <Skeleton width={160} height={160} />
-              )}
-              {!imageUploadState.loading && (imageUploadState.image || valueObject) && (
-                <img src={imageUploadState.image || valueObject} alt='image' loading='lazy' />
-              )}
-              <UploadImageIconContainer>
-                <UploadImageIcon>
-                  <BsCardImage />
-                  <span>{t('DRAG_AND_DROP', 'Drag and drop')}</span>
-                </UploadImageIcon>
-              </UploadImageIconContainer>
-            </DragAndDrop>
-          </ExamineClick>
-        </ComponentImage>
+        <ThemeImage
+          isMarginTop
+          valueObject={valueObject}
+          handleAddThemeGallery={handleAddThemeGallery}
+          handleChangeValue={handleChangeValue}
+        />
       )}
       {typeof optionObject !== 'string' && Object.keys(optionObject).filter(subOption => subOption !== 'value_type' && subOption !== 'options').map(subOption => (
         <React.Fragment key={subOption}>
@@ -213,28 +113,6 @@ export const ThemeOption = (props) => {
           )}
         </React.Fragment>
       ))}
-      <Alert
-        title={t('ORDERING_PRODUCTS', 'Ordering products')}
-        content={alertState.content}
-        acceptText={t('ACCEPT', 'Accept')}
-        open={alertState.open}
-        onClose={() => setAlertState({ open: false, content: [] })}
-        onAccept={() => setAlertState({ open: false, content: [] })}
-        closeOnBackdrop={false}
-      />
-      <Modal
-        width='700px'
-        height='80vh'
-        padding='30px'
-        title={t('IMAGE_CROP', 'Image crop')}
-        open={cropState?.open}
-        onClose={() => setCropState({ ...cropState, open: false })}
-      >
-        <ImageCrop
-          photo={cropState?.data}
-          handleChangePhoto={handleChangePhoto}
-        />
-      </Modal>
     </OptionContainer>
   )
 }
